+++ resolwe/flow/models/utils.py
@@ -346,6 +346,7 @@ def hydrate_input_references(input_, input_schema, hydrate_values=True):
                 output["__id"] = data.id
                 output["__type"] = data.process.type
                 output["__descriptor"] = data.descriptor
+                output["__name"] = getattr(data, "name", None)
                 output["__entity_name"] = getattr(data.entity, "name", None)
                 output["__output_schema"] = data.process.output_schema
 
@@ -370,6 +371,7 @@ def hydrate_input_references(input_, input_schema, hydrate_values=True):
                     output["__id"] = data.id
                     output["__type"] = data.process.type
                     output["__descriptor"] = data.descriptor
+                    output["__name"] = getattr(data, "name", None)
                     output["__entity_name"] = getattr(data.entity, "name", None)
                     output["__output_schema"] = data.process.output_schema